Updated plugins

+++ astra-sites/inc/lib/ai-builder/inc/classes/importer/class-ai-builder-importer.php
@@ -66,12 +66,14 @@ class Ai_Builder_Importer {
 	/**
 	 * Clear Cache.
 	 *
+	 * @return void
+	 *
 	 * @since  1.0.9
 	 */
 	public function clear_related_cache() {
 
 		// Clear 'Astra Addon' cache.
-		if ( is_callable( 'Astra_Minify::refresh_assets' ) ) {
+		if ( class_exists( 'Astra_Minify' ) && is_callable( 'Astra_Minify::refresh_assets' ) ) {
 			\Astra_Minify::refresh_assets();
 		}
 
@@ -99,17 +101,25 @@ class Ai_Builder_Importer {
 	/**
 	 * Delete related transients
 	 *
+	 * @return void
+	 *
 	 * @since 3.1.3
 	 */
 	public function delete_related_transient() {
 		delete_transient( 'astra_sites_batch_process_started' );
-		ST_Importer_File_System::get_instance()->delete_demo_content();
+
+		if ( class_exists( 'STImporter\Importer\ST_Importer_File_System' ) ) {
+			ST_Importer_File_System::get_instance()->delete_demo_content();
+		}
+
 		delete_option( 'ast_ai_import_current_url' );
 		delete_option( 'astra_sites_ai_import_started' );
 	}
 
 	/**
 	 * Include files.
+	 *
+	 * @return void
 	 *
 	 * @since  1.0.0
 	 */
@@ -124,6 +134,8 @@ class Ai_Builder_Importer {
 	/**
 	 * Get the API URL.
 	 *
+	 * @return string
+	 *
 	 * @since  1.0.0
 	 */
 	public static function get_api_domain() {
@@ -132,6 +144,8 @@ class Ai_Builder_Importer {
 
 	/**
 	 * Setter for $api_url
+	 *
+	 * @return void
 	 *
 	 * @since  1.0.0
 	 */
